% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StationData.R
\name{StationData}
\alias{StationData}
\title{Take an initial look at a new dataset, looking at name, class, and unique values for each column in a new data frame.}
\arguments{
\item{data_in}{A new data.frame. This should previously be loaded into the R environment.}
}
\value{
A list containing a snapshot of the new dataset. "Name" contains the name of each column, "Class" contains the class of each column, and "Unique" contains the unique values of each column. The list is called "Info" and will print when the function is executed.
}
\description{
Take an initial look at a new dataset, looking at name, class, and unique values for each column in a new data frame.
}
