% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_AHI_inputs.R
\name{Write_AHI_inputs}
\alias{Write_AHI_inputs}
\title{Write the AHI input data to an Excel workbook.}
\usage{
Write_AHI_inputs(input_data, n_frequency = 4, major_paths = NULL, outputPath)
}
\arguments{
\item{input_data}{The input avalanche record. This format must comply to the DHA avalanche database standards in order for the function to work.}

\item{n_frequency}{The minimum number of events required to include the avalanche path as a 'Major' avalanche path. Default is 4. Value will be ignored if 'major_paths' argument is specified.}

\item{major_paths}{Optional vector of path names, if you want to look at a specific list of paths.}

\item{outputPath}{Desired path to save Excel Workbook}
}
\value{
This will save the AHI data to a workbook that is similar to the format used by Hamre et al.
}
\description{
Write the AHI input data to an Excel workbook.
}
