% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AHI_template.R
\name{AHI_template}
\alias{AHI_template}
\title{Create the AHI template based on the excel workbook used by Hamre et al.}
\usage{
AHI_template(
  input_data,
  n_frequency = 1,
  major_paths = NULL,
  WADT = 7000,
  percent_cars = 0.98,
  length_car_m = 15,
  length_truck_m = 30,
  reaction_time = 2.5,
  speed_limit = 48,
  road_grade = -0.1,
  coeff_friction = 0.2,
  Q_cars_powder = 0,
  Q_trucks_powder = 0,
  Q_cars_light = 3,
  Q_trucks_light = 3,
  Q_cars_deep = 10,
  Q_trucks_deep = 10,
  Q_cars_plunging = 12,
  Q_trucks_plunging = 12,
  powder_RI = 100,
  light_RI = 100,
  deep_RI = 100,
  plunging_RI = 1000,
  path_lengths = NULL,
  wait_time = 1,
  Ps = 0.05,
  Ps_prime = 0.15,
  Lave_light = 0.3,
  Lave_deep = 0.7,
  Lave_plunging = 0.5
)
}
\arguments{
\item{input_data}{The input avalanche record. This format must comply to the DHA avalanche database PathNames in order for the function to work.}

\item{n_frequency}{The minimum number of events required to include the avalanche path as a 'Major' avalanche path. Default is 4. Value will be ignored if 'major_paths' argument is specified.}

\item{major_paths}{Optional vector of path names, if you want to look at a specific list of paths.}

\item{WADT}{Traffic volume?}

\item{percent_cars}{Percentage of traffic volume that is cars. Percentage of trucks calculated automatically based on this value (1-percent_cars).}

\item{length_car_m}{Average car length.}

\item{length_truck_m}{Average truck length.}

\item{reaction_time}{Driver's time to react in seconds. Used to calculate stopping distance. Default is 2.5.}

\item{speed_limit}{Speed limit in km/h. Default 48 km/h.}

\item{road_grade}{Grade of road in percentage as a decimal. Negative values indicate downhill slope. Default is -0.1.}

\item{coeff_friction}{The coefficient of friction between the tires and the road. It is typically assumed to be equal to 0.7 on a dry road and in the range from 0.3 to 0.4 on a wet road. Default is 0.2.}

\item{Q_cars_powder}{Q factor for cars for powder events. Default = 0.}

\item{Q_trucks_powder}{Q factor for trucks for powder events. Default = 0.}

\item{Q_cars_light}{Q factor for cars for light events. Default = 3.}

\item{Q_trucks_light}{Q factor for trucks for light events. Default = 3.}

\item{Q_cars_deep}{Q factor for cars for deep events. Default = 10.}

\item{Q_trucks_deep}{Q factor for trucks for deep events. Default = 10.}

\item{Q_cars_plunging}{Q factor for cars for plunging events. Default = 12.}

\item{Q_trucks_plunging}{Q factor for trucks for plunging events. Default = 12.}

\item{powder_RI}{Return Interval to use for powder events if no events occur in the record. Default is 100.}

\item{light_RI}{Return Interval to use for light events if no events occur in the record. Default is 100.}

\item{deep_RI}{Return Interval to use for deep events if no events occur in the record. Default is 100.}

\item{path_lengths}{Path to optional .csv with columns "path_count", "PathName",'length', "Jminus", and "Jplus" defining the road length of each path and space between previous and next path, in meters.}

\item{wait_time}{Time in hours for waiting traffic. Default = 1.}

\item{Ps}{Probability of a second avalanche running in path J with waiting traffic. Default 0.05 after Schaerer, 1989.}

\item{Ps_prime}{Probability of an avalanche releasing in path adjacent to J with waiting traffic. Default 0.15 after Schaerer, 1989.}

\item{Lave_light}{Adjustment estimate for average length on road for light events. Default 0.3 from Schaerer, 1989}

\item{Lave_deep}{Adjustment estimate for average length on road for deep events. Default 0.7 from Schaerer, 1989}

\item{Lave_plunging}{Adjustment estimate for average length on road for plunging events. Default 0.5 from Schaerer, 1989}

\item{plugning_RI}{Return Interval to use for deep events if no events occur in the record. Default is 1000.}
}
\value{
The tables used in the AHI template

\item{MajorPaths}{A character vector of the major paths.}
}
\description{
Create the AHI template based on the excel workbook used by Hamre et al.
}
