% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Write_AHI_template.R
\name{Write_AHI_template}
\alias{Write_AHI_template}
\title{Write the AHI input data to an Excel workbook.}
\usage{
Write_AHI_template(
  input_data,
  n_frequency = 1,
  major_paths = NULL,
  WADT = 7000,
  percent_cars = 0.98,
  length_car_m = 15,
  length_truck_m = 30,
  reaction_time = 2.5,
  speed_limit = 48,
  road_grade = -0.1,
  coeff_friction = 0.2,
  Q_cars_powder = 0,
  Q_trucks_powder = 0,
  Q_cars_light = 3,
  Q_trucks_light = 3,
  Q_cars_deep = 10,
  Q_trucks_deep = 10,
  Q_cars_plunging = 12,
  Q_trucks_plunging = 12,
  powder_RI = 100,
  light_RI = 100,
  deep_RI = 100,
  plunging_RI = 1000,
  path_info = NULL,
  outputPath,
  wait_time = 1,
  Ps = 0.05,
  Ps_prime = 0.3,
  Lave_light = 0.3,
  Lave_deep = 0.7,
  Lave_plunging = 0.5,
  missing_seasons = 0
)
}
\arguments{
\item{input_data}{The input avalanche record. This format must comply to the DHA avalanche database PathNames in order for the function to work.}

\item{n_frequency}{The minimum number of events required to include the avalanche path as a 'Major' avalanche path. Default is 4. Value will be ignored if 'major_paths' argument is specified.}

\item{major_paths}{Optional vector of path names, if you want to look at a specific list of paths.}

\item{WADT}{Traffic volume?}

\item{percent_cars}{percentage of traffic volume that is cars. percentage of trucks calcualted automatically based on this value (1-percent_cars)}

\item{length_car_m}{average car length}

\item{length_truck_m}{average truck length}

\item{reaction_time}{Driver's time to react in seconds. Used to calculate stopping distance. Default is 2.5.}

\item{speed_limit}{Speed limit in km/h. Default 48 km/h}

\item{road_grade}{Grade of road in percentage as a decimal. Negative values indicate downhill slope. Default is -0.1.}

\item{coeff_friction}{the coefficient of friction between the tires and the road. It is typically assumed to be equal to 0.7 on a dry road and in the range from 0.3 to 0.4 on a wet road. Default is 0.2.}

\item{Q_cars_powder}{Q factor for cars for powder events. Default = 0}

\item{Q_trucks_powder}{Q factor for trucks for powder events. Default = 0}

\item{Q_cars_light}{Q factor for cars for light events. Default = 3}

\item{Q_trucks_light}{Q factor for trucks for light events. Default = 3}

\item{Q_cars_deep}{Q factor for cars for deep events. Default = 10}

\item{Q_trucks_deep}{Q factor for trucks for deep events. Default = 10}

\item{Q_cars_plunging}{Q factor for cars for plunging events. Default = 12}

\item{Q_trucks_plunging}{Q factor for trucks for plunging events. Default = 12}

\item{powder_RI}{Return Interval to use for powder events if no events occur in the record. Default is 100.}

\item{light_RI}{Return Interval to use for light events if no events occur in the record. Default is 100.}

\item{deep_RI}{Return Interval to use for deep events if no events occur in the record. Default is 100.}

\item{plunging_RI}{Return Interval to use for deep events if no events occur in the record. Default is 1000.}

\item{path_info}{Optional data.frame with columns "path_count", "PathName",'length', "Jminus", "Jplus", and "missing_seasons", defining the road length of each path and space between previous and next path, in meters, and the number of seasons missing from the entire record for each path.}

\item{outputPath}{Desired path to save Excel Workbook}

\item{wait_time}{Time in hours for waiting traffic. Default = 1.}

\item{Ps}{Probability of a second avalanche running in path J with waiting traffic. Default 0.05 after Schaere, 1989.}

\item{Ps_prime}{Probability of an avalanche releasing in path adjacent to J with waiting traffic. Default 0.3 after Schaerer, 1989.}

\item{Lave_light}{Adjustment estimate for average length on road for light events. Default 0.3 from Schaerer, 1989}

\item{Lave_deep}{Adjustment estimate for average length on road for deep events. Default 0.7 from Schaerer, 1989}

\item{Lave_plunging}{Adjustment estimate for average length on road for plunging events. Default 0.5 from Schaerer, 1989}

\item{missing_seasons}{Optional number of seasons missing from record. Used to calculate Record Length for return intervals.}
}
\value{
This will save the AHI data to a workbook that is similar to the format used by Hamre et al.
}
\description{
Write the AHI input data to an Excel workbook.
}
